refactor. List String으로 바꿔주는 Util 생성 및 테스트 파일 디렉토리 정리
