+++ src/main/java/step3/utils/ConvertToStringFromList.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class ConvertToStringFromList {
-    public String convertToStringFromList(List<String> list) {
+    public static String convertToStringFromList(List<String> list) {
         String result = list.stream()
                 .map(String::valueOf)
                 .collect(Collectors.joining(", "));
