+++ src/main/java/step3/view/InputView.java
@@ -17,7 +17,6 @@ public class InputView {
         String inputCarNames = scan.next();
         carNames = InputValidator.validateCarName(inputCarNames);
 
-
         System.out.println("시도할 회수는 몇 회 인가요?");
         numberOfAttemps = scan.nextInt();
