--- src/test/java/step3/ConvertToStringFromListTest.java
@@ -1,4 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-class ConvertToStringFromListTest {
-  
-}
