+++ src/test/java/step3/controller/RacingGameTest.java
@@ -1,8 +1,7 @@
-package step3;
+package step3.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.controller.RacingGame;
 import step3.domain.Cars;
 import step3.utils.GameRule;
 import step3.utils.RandomUtil;
