+++ src/test/java/step3/domain/CarsTest.java
@@ -1,9 +1,7 @@
-package step3;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.Car;
-import step3.domain.Cars;
 
 import java.util.ArrayList;
 import java.util.List;
