+++ src/test/java/step3/utils/GameRuleTest.java
@@ -1,10 +1,7 @@
-package step3;
+package step3.utils;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.utils.GameRule;
-import step3.utils.NumberGenerator;
-import step3.utils.TestNumberGenerator;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
